refactor: move blocks to SoA
## Summary

Refactors RB-CSR edge blocks from array-of-structs storage to struct-of-arrays storage.

## Changes

- Replace forward edge blocks with SoA columns for destinations, relations, and flags.
- Replace reverse edge blocks with a source-id array.
- Move per-block liveness from in-block masks to live-count sidecars.
- Add `src/storage/edge_blocks.zig` as the accessor layer for block reads/writes.
- Update adjacency traversal, mutation, repair, validation, and snapshot paths to use the new layout.
- Keep forward destination scans cache-friendly by reading a contiguous `u32` destination column.
- Align block payloads for cache-line-friendly scans.
- Update tests around storage accessors, adjacency runs, live counts, validation, mutation, and repair.
- Update README implementation notes to describe the SoA block layout.

## Why

The old AoS layout made neighbor scans touch relation/flag data even when only destination/source ids were needed. The SoA layout keeps hot traversal data contiguous and isolates layout details behind accessors so future storage changes are cheaper.

+++ src/query/snapshot/view.zig
@@ -11,7 +11,7 @@ pub const SnapshotSide = snapshot_capture.SnapshotSide;
 
 pub const CapturedGraphView = struct {
     core: *const graph_core.GraphCore,
-    node_state: []u32,
+    node_state: []u8,
     fwd_first_block: []u32,
     fwd_block_count: []u32,
     fwd_group_count: []u16,
